Extract alert to file
