+++ Blockzilla/BrowserViewController.swift
@@ -1920,6 +1920,8 @@ extension BrowserViewController: MenuActionable {
     }
 }
 
+// MARK: - Shortcuts Actions
+
 extension BrowserViewController {
     func showRenameAlert(shortcut: Shortcut) {
         let alert = UIAlertController.renameAlertController(
@@ -1952,32 +1954,7 @@ extension BrowserViewController {
         submit(url: url)
         GleanMetrics.Shortcuts.shortcutOpenedCounter.add()
     }
-}
 
-extension UIAlertController {
-    static func renameAlertController(currentName: String, renameAction: @escaping (_ newName: String) -> Void, cancelAction: @escaping () -> Void) -> UIAlertController {
-        let alert = UIAlertController(title: UIConstants.strings.renameShortcut, message: nil, preferredStyle: .alert)
-        alert.addTextField { textfield in
-            textfield.placeholder = UIConstants.strings.renameShortcutAlertPlaceholder
-            textfield.text = currentName
-            textfield.clearButtonMode = .always
-            NotificationCenter.default.addObserver(forName: UITextField.textDidChangeNotification, object: textfield, queue: OperationQueue.main, using: { _ in
-                alert.actions.last?.isEnabled = !(textfield.text?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty ?? false)
-            })
-        }
-
-        alert.addAction(UIAlertAction(title: UIConstants.strings.renameShortcutAlertSecondaryAction, style: .cancel, handler: { _ in
-            cancelAction()
-        }))
-        alert.addAction(UIAlertAction(title: UIConstants.strings.renameShortcutAlertPrimaryAction, style: .default, handler: { [unowned alert] action in
-            let newName = (alert.textFields?.first?.text ?? currentName).trimmingCharacters(in: .whitespacesAndNewlines)
-            renameAction(newName)
-        }))
-        return alert
-    }
-}
-
-extension BrowserViewController {
     func shortcutDidRemove(shortcut: ShortcutViewModel) {
         for subview in shortcutsContainer.subviews where subview is ShortcutView {
             let subview = subview as? ShortcutView
